requirements-dev.txt
asgiref==3.2.3
Django==3.0.3
django-cors-headers==3.2.1
django-environ==0.4.5
djangorestframework==3.11.0
pip-autoremove==0.9.1
pkg-resources==0.0.0
psycopg2-binary==2.8.4
pytz==2019.3
sqlparse==0.3.0
